sha1: moved self test to separated test
