+++ .gitignore
@@ -48,6 +48,7 @@ tests/test_jid
 tests/test_rand
 tests/test_sasl
 tests/test_scram
+tests/test_sha1
 tests/test_sock
 m4/
 libstrophe.project
