+++ Makefile.am
@@ -65,7 +65,8 @@ examples_uuid_LDADD = $(STROPHE_LIBS)
 
 
 ## Tests
-TESTS = tests/check_parser tests/test_rand tests/test_scram tests/test_base64
+TESTS = tests/check_parser tests/test_sha1 tests/test_rand tests/test_scram \
+	tests/test_base64
 check_PROGRAMS = $(TESTS)
 
 tests_check_parser_SOURCES = tests/check_parser.c tests/test.h
@@ -84,3 +85,6 @@ tests_test_rand_CFLAGS = $(STROPHE_FLAGS) -I$(top_srcdir)/src
 
 tests_test_scram_SOURCES = tests/test_scram.c tests/test.c src/sha1.c
 tests_test_scram_CFLAGS = $(STROPHE_FLAGS) -I$(top_srcdir)/src
+
+tests_test_sha1_SOURCES = tests/test_sha1.c src/sha1.c
+tests_test_sha1_CFLAGS = -I$(top_srcdir)/src
